+++ src/NvimClient.API/NvimUnhandledRequestEventArgs.cs
@@ -6,7 +6,7 @@ namespace NvimClient.API
   {
     private readonly NvimAPI _nvim;
 
-    internal NvimUnhandledRequestEventArgs(NvimAPI nvim, long requestId,
+    internal NvimUnhandledRequestEventArgs(NvimAPI nvim, uint requestId,
       string methodName, object[] arguments)
     {
       _nvim      = nvim;
@@ -17,7 +17,7 @@ namespace NvimClient.API
 
     public string   MethodName { get; }
     public object[] Arguments  { get; }
-    public long     RequestId  { get; }
+    public uint     RequestId  { get; }
 
     public void SendResponse(object result, object error = null) =>
       _nvim.SendResponse(this, result, error);
