+++ src/NvimClient/NvimMsgpack/NvimMessageTypeAttribute.cs
@@ -5,7 +5,7 @@ namespace NvimClient.NvimMsgpack
   [AttributeUsage(AttributeTargets.Class)]
   internal class NvimMessageTypeAttribute : Attribute
   {
-    public NvimMessageTypeAttribute(long id) => Id = id;
-    public long Id { get; }
+    public NvimMessageTypeAttribute(byte id) => Id = id;
+    public byte Id { get; }
   }
 }
